Fix lighting seams, replace models
Finally a decent lighting build! The seams problem was caused because the
segments of wall had interior faces, changing the models so that parts
between segments don't have faces fixed the problem. As such, there are
now separate models for mid wall and end wall segments, as well as some
new corner types of wall. The level has been rebuilt, granted it still is
just a sample level for testing the visuals.